change int to qlonglong

+++ core/libs/database/coredb/coredb.cpp
@@ -596,8 +596,8 @@ void CoreDB::deleteTag(int tagID)
 
 void CoreDB::setTagIcon(int tagID, const QString& iconKDE, qlonglong iconID)
 {
-    int     dbIconID  = iconKDE.isEmpty() ? iconID : 0;
-    QString dbIconKDE = iconKDE;
+    qlonglong dbIconID = iconKDE.isEmpty() ? iconID : 0;
+    QString dbIconKDE  = iconKDE;
 
     if (iconKDE.isEmpty()                            ||
         (iconKDE.toLower() == QLatin1String("tag"))  ||
